% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider.R
\name{add_queue}
\alias{add_queue}
\title{Add items to a spider queue}
\usage{
add_queue(.x, items)
}
\description{
Add items to a spider queue
}
\examples{
spider() \%>\% add_queue("first item")

spider() \%>\% add_queue(c("first item", "second item"))
}
\seealso{
Other helpers: 
\code{\link{add_step}()},
\code{\link{run}()},
\code{\link{set_name}()},
\code{\link{set_pipeline}()}
}
\concept{helpers}
