% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{add_step}
\alias{add_step}
\alias{add_parser}
\alias{add_transformer}
\title{Add steps to a pipeline or spider.}
\usage{
add_step(.x, step, after = length(.x$steps))

add_parser(.x, ..., after = length(.x$steps))

add_transformer(.x, ..., after = length(.x$steps))
}
\arguments{
\item{.x}{the spider or pipeline to update}

\item{step}{the parser or transformer step to add}

\item{after}{the index after which the step should be added}
}
\description{
adds a step to the given pipeline.
}
\examples{
spider() \%>\% add_parser(~ .x * 2)

spider() \%>\% add_parser(~ .x * 2, name = "multiply by 2")

spider() \%>\% add_transformer(~ sum(.x))

spider() \%>\% add_transformer(~ sum(.x)), name = "get the sum")
}
\seealso{
Other helpers: 
\code{\link{add_queue}()},
\code{\link{run}()},
\code{\link{set_name}()},
\code{\link{set_pipeline}()}
}
\concept{helpers}
