% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{add_transformer}
\alias{add_transformer}
\title{add_transformer}
\usage{
add_transformer(.x, parser)
}
\description{
adds a transformer step to the given pipeline.
Transformers are applied across all items of the pipeline and can often be used for aggregation.
}
