% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{d}
\alias{d}
\title{debug
sets a debugger breakpoint}
\usage{
d(obj, head = 1, parallel = F)
}
\description{
This will automatically start running the scraper and drop into the debugger where placed. It assigns environment variables .x as the first item in the queue and .d as the entire queue.
}
