% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.r
\name{login}
\alias{login}
\title{Handle the login page
 - Close the cookies banner
 - Enter username and password
 - Wait for the transactions page to load}
\usage{
login(client, username, password)
}
\arguments{
\item{client}{}

\item{username}{}

\item{password}{}
}
\description{
Handle the login page
 - Close the cookies banner
 - Enter username and password
 - Wait for the transactions page to load
}
