% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeley-local-scrape.R
\name{Mendeley_DF}
\alias{Mendeley_DF}
\title{Scrape information from a mendeley sqlite local database}
\usage{

  Mendeley_DF(path = "C:/Users/Oliver/AppData/Local/Mendeley Ltd/Mendeley Desktop/oj.watson@hotmail.co.uk@www.mendeley.com.sqlite")
}
\arguments{
\item{path}{File path for local mendeley database. Usually found in users/x/AppData/Local/Mendeley Ltd/Mendeley Desktop/...sqlite}
}
\value{
dataframe of scraped information relating to mendeley database
}
\description{
\code{Mendeley_DF} takes a provided file path for the location of your local mendeley sqlite database.
Function ultimately links with the shiny directory (to be moved to own package soon) to help with
visualisation and semantic creation of useful literature tool.
}

