% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Collection of generic steps to append to spider}
\usage{
pipeline(name = "", steps = list())
}
\arguments{
\item{name}{name of the pipeline}

\item{steps}{a \code{\link[=list]{list()}} of \code{\link[=steps]{steps()}} to be applied}
}
\description{
combination of multiple steps to be applied in parellel to data
}
\examples{
pipeline("clean names") |>
    add_transformer(
        ~ janitor::clean_names(.x)
    )
}
