% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{Post_Process_Recipes}
\alias{Post_Process_Recipes}
\title{BBC Food recipes post processing}
\usage{
Post_Process_Recipes(recipe.list)
}
\description{
\code{Post_Process_Recipes} takes the sraped recipe list from \code{Recipe_Pull} and cleans text encode errors
}

