% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe-pull.R
\name{Recipe_Pull}
\alias{Recipe_Pull}
\title{BBC Food recipes scrape}
\usage{
Recipe_Pull(sitemap.url = NULL, sitemap.path = NULL, failed = NULL)
}
\arguments{
\item{sitemap.url}{URL of sitemap. Default = NULL which triggers reading a sitemap listed in sitemap.path. If sitemap.path is
not NULL then this will be used in stead of sitemap.url}

\item{sitemap.path}{Path of sitemap file. Default = NULL which triggers reading the saved bbc.co.uk/food/sitemap.xml data.}

\item{failed}{Vector which is returned from \code{Recipe_Pull} which contains the positions of failed urls}
}
\value{
List of list of recipes
}
\description{
\code{Recipe_Pull} takes the saved sitemap of the bccfood website and uses it to pull all the
recipes from the website into a list of lists for futher use.
}

