% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pipeline.R, R/runner.R,
%   R/spider.R, R/steps.R
\name{run}
\alias{run}
\alias{run.pipeline}
\alias{run.runner}
\alias{run.spider}
\alias{run.parser}
\alias{run.transformer}
\title{Run a spider or a runner}
\usage{
run(.x, ...)

\method{run}{pipeline}(.x, .data)

\method{run}{runner}(.x)

\method{run}{spider}(.x)

\method{run}{parser}(.x, .data, spider = spider())

\method{run}{transformer}(.x, .data, spider = spider())
}
\description{
The run function goes through each step of a spider's steps and then runs each of the steps in it's pipeline.
}
\seealso{
Other helpers: 
\code{\link{add_queue}()},
\code{\link{add_step}()},
\code{\link{set_name}()},
\code{\link{set_pipeline}()}
}
\concept{helpers}
