% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape}
\alias{scrape}
\title{Metadata in documents}
\usage{
scrape(wd = NULL, filetype = NULL, verbose = F, cutoff = 1)
}
\arguments{
\item{wd}{A file directory, defaults to current working directory}

\item{filetype}{The type of files to scrape, defaults to pdf only}

\item{cutoff}{The proportion of documents with metadata field to return, defaults to 1 (all fields)}
}
\value{
A dataframe with the metadata for all documents within the
file path directory
}
\description{
Metadata in documents
}
\examples{
scrape()
}
