% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datepicker.r
\name{select_all}
\alias{select_all}
\title{Select all dates from an initial date to today's date
 - Open the date widget
 - Navigate to the start month
 - Click the start day
 - Navigate to the current month
 - Click today's day
 - Submit the form}
\usage{
select_all(client, start_date)
}
\arguments{
\item{client}{}

\item{start_date}{}
}
\description{
Select all dates from an initial date to today's date
 - Open the date widget
 - Navigate to the start month
 - Click the start day
 - Navigate to the current month
 - Click today's day
 - Submit the form
}
