% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider.R
\name{set_name}
\alias{set_name}
\title{Rename a spider}
\usage{
set_name(.x, name)
}
\description{
Redefine the name of a spider
}
\examples{
# this is often used after creating a template spider

template <- spider("template")

implementation_one <- set_name(spider, "implementation_one")
implementation_two <- set_name(spider, "implementation_two")
}
\seealso{
Other helpers: 
\code{\link{add_step}()},
\code{\link{run}()},
\code{\link{set_pipeline}()}
}
\concept{helpers}
