% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_pipeline}
\alias{set_pipeline}
\title{pipeline
sets the pipeline used by the given spider}
\usage{
set_pipeline(.x, parser)
}
\description{
pipeline
sets the pipeline used by the given spider
}
\seealso{
Other helpers: 
\code{\link{add_step}()},
\code{\link{run}()},
\code{\link{set_name}()}
}
\concept{helpers}
