% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider.R
\name{spider}
\alias{spider}
\title{Spider for crawling urls}
\usage{
spider(name = "", queue = list(), steps = list(), pipeline = NA)
}
\arguments{
\item{name}{the name of the spider to be used for logging purposes}

\item{queue}{the \code{\link[=list]{list()}} of urls or initial paths to be used by \code{\link[=steps]{steps()}}}

\item{steps}{a \code{\link[=list]{list()}} of \code{\link[=steps]{steps()}} to be applied to the data sequencially}

\item{pipeline}{the default \code{\link[=pipeline]{pipeline()}} to apply to the data after all steps are applied}
}
\description{
combination of multiple steps to be applied in parellel to data
}
\examples{
s_multiply <- spider('multiply', queue = 1:4, steps = list(parser( ~ .x * 2)))
s_multiply

run(s_multiply)
}
