% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{steps}
\alias{steps}
\alias{parser}
\alias{transformer}
\title{Steps for Spiders and Pipelines}
\usage{
parser(.f, name = paste(.f, collapse = ""), flatten = T, apply = NA)

transformer(.f, name = paste(.f, collapse = ""))
}
\arguments{
\item{.f}{formula to be used in the step that accepts the input .x and optional .y parameters}

\item{name}{The name of the object}

\item{flatten}{whether or not to flatten the result of the parser}

\item{apply}{a function to apply to your parser function if you don't want to use \code{\link[purrr:map]{purrr::map()}}}
}
\description{
Steps to be added to either the pipeline steps or spider steps.
}
\examples{
p_mult_2 <- parser( ~ .x * 2, name = "multply by two")
p_mult_2

run(p_mult_2, 1:5)

t_sum <- transformer(~ sum(.x), name = "get sum")
t_sum

run(t_sum, 1:5)
}
\seealso{
\code{\link[=add_step]{add_step()}} \code{\link[=add_parser]{add_parser()}} \code{\link[=add_transformer]{add_transformer()}} \code{\link[=pipeline]{pipeline()}} \code{\link[=spider]{spider()}}
}
