% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{t_clean_names}
\alias{t_clean_names}
\title{clean names}
\usage{
t_clean_names()
}
\value{
\code{\link[=transformer]{transformer()}}
}
\description{
Runs the janiotr clean_names function on the queue.
}
