% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{t_save_aws}
\alias{t_save_aws}
\title{write results to aws s3}
\usage{
t_save_aws(
  bucket = getOption("aws_bucket"),
  dir = getOption("aws_dir", "data")
)
}
\arguments{
\item{bucket}{The aws bucket to save to. Can be set globally with \code{\link[=options]{options()}}}

\item{dir}{The directory in the bucket to save to. Can be set globally with \code{\link[=options]{options()}}}
}
\description{
To use this function, make sure that you have set the following in your .Renviron:
\itemize{
\item AWS_ACCESS_KEY_ID
\item AWS_SECRET_ACCESS_KEY
\item AWS_DEFAULT_REGION
}
}
