% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{t_save_output}
\alias{t_save_output}
\title{save output}
\usage{
t_save_output(dir = getOption("output_dir", "output"))
}
\arguments{
\item{dir}{the directory to save the csv to. Can be set globally with \code{\link[=options]{options()}}}
}
\value{
\code{\link[=transformer]{transformer()}}
}
\description{
Saves the current queue as a csv file in the directory of your choice.
The name of the csv file will match the name of the spider.
}
