% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wayback-pull.R
\name{Wayback_Pull}
\alias{Wayback_Pull}
\title{Wayback machine scrape example for table information}
\usage{

  Wayback_Pull(url.vector = c("https://web.archive.org/web/20150401000000*/https://www.fxcm.com/forex/forex-pricing/",
  "https://web.archive.org/web/20160901000000*/https://www.fxcm.com/forex/forex-pricing/"),
  css = ".center td", col.names = c("Pair", "Spread"))
}
\arguments{
\item{url.vector}{Vector of urls from wayback machine archives}

\item{css}{Minimal css selector for a table of information}
}
\value{
list of dataframes of scraped information
}
\description{
\code{Wayback_Pull} takes a list of urls from the archive pages for your url of interest (See default url.vector
argument for an example) and scrapes from your wanted urls a minimum css selector based table. Somewhat arbritray
based on table structure of this page but could easily be extended to loop over css selection for table columns etc.
}

