% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{add_step}
\alias{add_step}
\alias{add_parser}
\alias{add_transformer}
\title{Add steps to a pipeline or spider.}
\usage{
add_step(.x, parser)

add_parser(.x, parser)

add_transformer(.x, parser)
}
\description{
adds a step to the given pipeline.
}
\seealso{
Other helpers: 
\code{\link{run}()},
\code{\link{set_name}()},
\code{\link{set_pipeline}()}
}
\concept{helpers}
