% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Collection of generic steps to append to spider}
\usage{
pipeline(name = "", steps = list())
}
\description{
combination of multiple steps to be applied in parellel to data
}
\examples{
pipeline("clean names") |>
    add_transformer(
        ~ janitor::clean_names(.x)
    )
}
