% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe-pull.R
\name{Recipe_Pull}
\alias{Recipe_Pull}
\title{BBC Food recipes scrape}
\usage{
Recipe_Pull(sitemap = NULL, start = 15, end = 11141)
}
\arguments{
\item{sitemap}{url of sitemap. Defauly NULL but in theory could specify url of recipe site sitemap that uses
recipe schema}

\item{start}{Start of recipe urls in sitemap data}

\item{end}{End of recipe urls in sitemap data}
}
\value{
List of list of recipes
}
\description{
\code{Recipe_Pull} takes the saved sitemap of the bccfood website and uses it to pull all the
recipes from the website into a list of lists for futher use.
}

