% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{run}
\alias{run}
\title{Run a spider or a runner}
\usage{
run(.x)
}
\description{
The run function goes through each step of a spider's steps and then runs each of the steps in it's pipeline.
}
\seealso{
Other helpers: 
\code{\link{add_step}()},
\code{\link{set_name}()},
\code{\link{set_pipeline}()}
}
\concept{helpers}
