\name{scrapeR-package}
\Rdversion{1.1}
\alias{scrapeR-package}
\docType{package}
\title{
Tools for Scraping Data from Web-Based Documents
}
\description{
Tools for Scraping Data from Web-Based Documents
}
\details{
\tabular{ll}{
Package: \tab scrapeR\cr
Type: \tab Package\cr
Version: \tab 0.1.5\cr
Date: \tab 2009-11-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Ryan M. Acton <rmacton@gmail.com> \url{http://www.ryanacton.com}}

\references{
Duncan Temple Lang. (2009). XML: Tools for parsing and generating XML within R and S-Plus. \url{http://CRAN.R-project.org/package=XML}.

Duncan Temple Lang. (2009). RCurl: General network (HTTP/FTP/...) client interface for R. \url{http://CRAN.R-project.org/package=RCurl}.
}
\keyword{ package }
