% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{waities.package}
\alias{waities.package}
\title{waities: Website scraping and curation tools for BBC food backup}
\description{
waities: Website scraping and curation tools for BBC food backup
}
\author{
OJ Watson \email{oj.watson@hotmail.co.uk}
}

